Pub/Sub sensor/carpark Initialise
Setup the pub/sub communication for the simple_mqtt_sensor and simple_mqtt_carpark. Sensor now takes keyboard inputs to replicate a car entering and leaving and the carpark will display the time, the spaces and the temperature when it receives an input. These values are still default values and will be updated later.